Move the ipython processor into the magics module, remove from defaults
