+++ pyinstrument/renderers/console.py
@@ -280,7 +280,6 @@ class ConsoleRenderer(FrameRenderer):
             processors.remove_irrelevant_nodes,
             processors.remove_first_pyinstrument_frames_processor,
             processors.group_library_frames_processor,
-            processors.strip_ipython_frames_processor,
         ]
 
     class colors_enabled:
