+++ pyinstrument/renderers/html.py
@@ -111,5 +111,4 @@ class HTMLRenderer(FrameRenderer):
             processors.remove_irrelevant_nodes,
             processors.remove_first_pyinstrument_frames_processor,
             processors.group_library_frames_processor,
-            processors.strip_ipython_frames_processor,
         ]
